Knight Horse Rider
🚀 Just conquered the Knight's Tour problem on GeeksforGeeks! 🎯

It was a challenging yet rewarding experience navigating through the intricacies of backtracking and recursion. The satisfaction of finding a solution that allows the knight to touch every square on the chessboard without repeating is unparalleled! 🧠✨

This problem not only sharpened my problem-solving skills but also deepened my understanding of algorithmic logic. 💻🔍

#CodingJourney #AlgorithmChallenges #GeeksforGeeks #ProblemSolving #ChessLogic #Backtracking